pyproject.toml
[project]
name = "python-crypto-trading-bot"
version = "0.1.0"
description = "Fools Flagon 2026 Crypto Trading Bot Experiment"
authors = [
    {name = "Fools Flagon"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"

# Runtime dependencies
dependencies = [
    "sqlalchemy",
    "pydantic",
    "python-dotenv",
    "python-dateutil",
    "pyyaml",
    "psycopg[binary] (>=3.2.12,<4.0.0)",
    "coinbase-advanced-py (>=1.8.2,<2.0.0)",
]

# Poetry-specific configuration
[tool.poetry]
# Tell Poetry what folder contains your package code
packages = [
  { include = "bot" }
]

# Development dependencies (only installed when running locally)
[tool.poetry.group.dev.dependencies]
alembic = "*"
black = "*"
isort = "*"
mypy = "*"

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"
